apps/voice-server/pyproject.toml
[project]
name = "voice-server"
version = "0.1.0"
description = "Smart Choice Auto Shop Voice Agent - Pipecat + Gemini Live API"
requires-python = ">=3.10"
dependencies = [
    "pipecat-ai[google,daily,silero]>=0.0.108",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "asyncpg>=0.29.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.0.0",
    "twilio>=9.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
